Remove Scala.js from validation project
