+++ validation/src/test/scala/hmda/validation/rules/ts/syntactical/S100Spec.scala
